Drop All Values After The Tag When Parsing
Coursier Version parses "0.0.0.1-RC1" as `Vector(0, 0, 0, 1 RC, 1)`. We want to drop all the values after the tag, but also record if the version was tagged.

The git automatic previous version code was also changed to attempt to find the first _non-tagged_ (in the version sense) tag (in the git sense), and then fallback to the first tagged version. This is because when working with some version "0.0.0.2-RC2" users are likely interested comparing to "0.0.0.1" _not_ "0.0.0.2-RC1".