Allow versionSchemeEnforcerPreviousVersion To Be Empty
When versionSchemeEnforcerPreviousVersion is empty and versionSchemeCheck is run, this no longer causes a failure. Now versionSchemeEnforcerPreviousVersion being empty can be used to denote that there is no previous version and thus binary compatibility checks should not be run.

This is useful when adding a new module to an existing project.