Add Support For Looking Up Multiple Tags From VCS
This commit adds support for looking up multiple tags from supported VCS systems. It also adds a basic framework for adding more VCS systems in the future. Finally, it adds some tests for the VCS systems which were difficult before since it is hard to test a clean git repo inside a git repo.

This commit does not change the functionality of the plugin at this time.