Add The Ability To Filter Out Certain VCS Tags
Often when selecting a tag to calculate the previous version we want to ignore certain tags. A common example of this is milestone releases. When comparing the binary compatibility of `1.1.0.0-M2` to a previous release we likely want to select the most recent _non milestone_ release.

This commit adds support for this by adding the `versionSchemeEnforcerPreviousTagFilter` key.