Add Concept Of Tag Metadata
This commit adds the concept of a VCS Tag as a data type. This allows us to model associated tag metadata, e.g. the creation time of the tag. This will become useful as we perform more complex tag detection, e.g. based on the most recent date and/or the most local binary change in terms of the version number and scheme.

This introduces two new setting keys,

* `versionSchemeEnforcerPreviousVCSTagFilter`
* `versionSchemeEnforcerPreviousVCSTagStringFilter`

These replace `versionSchemeEnforcerPreviousTagFilter` which is now deprecated.

Because of some code restructuring we now record the version suffix in `versionSchemeEnforcerPreviousVersion` if it is detected automatically, e.g. `0.0.0.1-M1` doesn't get truncated to `0.0.0.1` at this stage. This is good, as we will eventually want to allow for more complex actions to be take differentiating on that state, but this commit doesn't actually change the underlying behavior yet, we still eventually remove the suffix, just at a later stage.

The `VCS` type class was update substantially. Most of the changes involve returning `Tag` types now instead of `String` types. Additionally many methods return a `SortedSet[Tag]` rather than a `Stream`. This is because future changes may want to discriminate on the entire set of tags at once. Finally, the most outer signature of most methods has changed to be wrapped in an `Either[Throwable, *]`. This is because the invocation of the VCS system to read tags may fail. Before we were just treating that as an empty list of tags, but that is not really valid. As of now we don't report this error to the user, though we should, but that will require some additional restructuring.