Add Fundamental Data Types For Versions
This commit adds 7 data types which are the fundamental data types for parsing and representing most version strings in most version schemes.

Note, there are likely some changes that should come in soon after this to other existing types, but are not included in this commit.

Note, all the examples here assume PVP or SemVer.

Token Types
==============

* NumericVersionToken
    * Represents individual tokens from the numeric section of a version string.
    * For example, in the version string "1.0.0", there are three `NumericVersionToken` values, 1, 0, and 0.
* PreReleaseVersionToken
    * Represents individual tokens from the pre-release section of a version string.
    * For example, in the version string "1.0.0-SNAPSHOT", the `PreReleaseVersionToken` value would be `SNAPSHOT`.
* MetadataVersionToken
    * Represents individual tokens from the metadata section of a version string.
    * For example, in the version string "1.0.0-SNAPSHOT+b6b6d99", the `MetadataVersionToken` value would be `b6b6d99`.

Section Types
=================

Section types are composed of zero or more token values.

* NumericSection
    * Represents the set of `NumericVersionToken` values which make up the numeric section of a version string.
    * For example, in the version string "1.0.0", the `NumericSection` is `1.0.0`.
* PreReleaseSection
    * Represents the set of `PreReleaseVersionToken` values which make up the pre-release section of a version string.
    * For example, in the version string "1.0.0-RC.1", the `PreReleaseSection` is `-RC.1`.
* MetadataSection
    * Represents the set of `MetadataVersionToken` values which make up the metadata section of a version string.
    * For example, in the version string "1.0.0-RC.1+b6b6d99", the `MetadataSection` is `b6b6d99`.

VersionSections
====================

The `VersionSections` type represents an entire version string, composed of a `NumericSection` (which may be empty), an optional `PreReleaseSection`, and an optional `Metadata`.

It can be used to create and operate on version data types adhering to a particular version scheme. Note, these data types are not yet present in this library.